MAT-6851 added test case
